% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{print.msg}
\alias{print.msg}
\title{Pretty print a 'msg' object}
\usage{
\method{print}{msg}(x, ...)
}
\arguments{
\item{x}{a "`msg`" object}

\item{...}{unused}
}
\description{
Pretty print a 'msg' object
}
