% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-msg.r
\name{read_msg}
\alias{read_msg}
\title{Read an Outlook '.msg' file into a data frame}
\usage{
read_msg(path)
}
\arguments{
\item{path}{path to '.msg' file}
}
\value{
data frame
}
\description{
Read an Outlook '.msg' file into a data frame
}
\examples{
read_msg(system.file("extdata/unicode.msg", package="msgxtractr"))
}
