% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tibble.R
\name{tidy_msg}
\alias{tidy_msg}
\alias{as_tibble.msg}
\title{Turn a single `msg` object into a `tibble`}
\usage{
tidy_msg(x, ...)

\method{as_tibble}{msg}(x, ...)
}
\arguments{
\item{x}{a `msg` object}

\item{...}{passed on to [tibble::as_tibble()]}
}
\description{
Turn a single `msg` object into a `tibble`
}
