% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgxtractr-package.R
\docType{package}
\name{msgxtractr}
\alias{msgxtractr-package}
\alias{msgxtractr}
\title{Read Outlook '.msg' Files}
\description{
'Microsoft' 'Outlook' messages can be saved in '.msg' files. Tools are provided that
enable extraction of metadata, envelope, headers, body and attachments from these
files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/certe-medical-epidemiology/msgxtractr}
  \item Report bugs at \url{https://github.com/certe-medical-epidemiology/msgxtractr/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
