% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-msg.r
\name{read_msg}
\alias{read_msg}
\title{Read in an Outlook '.msg' file}
\usage{
read_msg(path)
}
\arguments{
\item{path}{path to '.msg' file}
}
\value{
a \code{list} of extracted fields and metadata
}
\description{
Read in an Outlook '.msg' file
}
\examples{
read_msg(system.file("extdata/unicode.msg", package="msgxtractr"))
read_msg(system.file("extdata/TestMessage-ansi.msg", package="msgxtractr"))
read_msg(system.file("extdata/TestMessage-default.msg", package="msgxtractr"))
read_msg(system.file("extdata/TestMessage-unicode.msg", package="msgxtractr"))
}
