% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.r
\name{save_attachments}
\alias{save_attachments}
\title{Save all attachments from a 'msg" object}
\usage{
save_attachments(msg_obj, path = getwd(), use_short = TRUE,
  quiet = FALSE)
}
\arguments{
\item{msg_obj}{a message object read in with \code{read_msg()}}

\item{path}{directory path to save attachments in (defaults to current directory)}

\item{use_short}{if \code{TRUE} then use the "short" filename from the parsed message,
otherwise use the "long" filename (if it exists)}

\item{quiet}{if \code{TRUE} then no informative messages will be displayed}
}
\value{
a character vector of full path names of files written out (invisibly)
}
\description{
Save all attachments from a 'msg" object
}
\examples{
x <- read_msg(system.file("extdata/unicode.msg", package="msgxtractr"))
td <- tempdir()
res <- save_attachments(x, td)
for (f in res) unlink(f)
}
